+++ test/test_commons.py
@@ -9,7 +9,6 @@ import typing
 import pytest
 
 import technologydata
-from technologydata import Commons
 
 
 class TestCommonsUtils:
@@ -169,18 +168,3 @@ class TestCommonsUtils:
             input_unit, input_year
         )
         assert result == expected_string
-
-    def test_defaults_and_flag(self, monkeypatch: typing.Any) -> None:
-        """Check if parse_input_arguments works when both default arg and the store_true flag are provided."""
-        monkeypatch.setattr("sys.argv", ["prog", "--num_digits", "2", "--store_source"])
-        args = Commons.parse_input_arguments()
-        assert args.num_digits == 2
-        assert args.store_source is True
-
-    def test_default_values_when_not_provided(self, monkeypatch: typing.Any) -> None:
-        """Check if parse_input_arguments works when no args are provided (so defaults are used)."""
-        # no args -> defaults apply
-        monkeypatch.setattr("sys.argv", ["prog"])
-        args = Commons.parse_input_arguments()
-        assert args.num_digits == 4  # default from the function
-        assert args.store_source is False
